requirements.txt
requests
urllib3
dnspython
netaddr